Swapped to device service style.

+++ HomeHook.Common/Models/MediaItem.cs
@@ -8,7 +8,7 @@
         public required MediaMetadata Metadata { get; set; }
         public string User { get; set; } = string.Empty;
         public bool Cache { get; set; } = false;
-        public float StartTime { get; set; } = 0;
-        public float Runtime { get; set; } = 0;
+        public double StartTime { get; set; } = 0;
+        public double Runtime { get; set; } = 0;
     }
 }
